+++ json-circe/shared/src/test/scala/busymachines/pureharm/json/test/derivetest/JsonDefaultSemiAutoDecoderDerivationTest.scala
@@ -56,7 +56,7 @@ final class JsonDefaultSemiAutoDecoderDerivationTest extends JsonTest {
       val errors = compileErrors(
         """
           val rawJson = "{}"
-          rawJson.unsafeDecodeAs[WinterMelon]
+          rawJson.decodeAs[WinterMelon]
          """
       )
       assert(